Exceptions thrown if Events are out of order
There are two ensure event order functions made available for the main
pipeline.

The first, and curretly used, function throws an exception when an out
of order event is discovered. We don't know exactly how bad this issue
will be, but this will prevent us experiencing the nasty effects when
it crops up.

The second is a suspected solution to the problem. The implementation
appears to work, extensive time has been with it, but writing a
function capable of truely determining it in the generator framework
is taking too long. Therefore, given that we don't expect this to
occur for some time and we will be alerted to an instance, it's being
moth balled here until it's needed.